+++ src/utils/serialiseNodesEdges/showEdgeCheck.js
@@ -1,11 +1,23 @@
+/**
+ * Check if edge is displayable
+ * @param  {Object}   params
+ * @param  {Object}   params.addedEdges              Array of edges IDs been added
+ * @param  {Object}   params.edgeId                  Edge ID
+ * @param  {Array}    params.edgesIdsToDisplay       Array of edges IDs to display
+ * @param  {String}   params.edgeUniqueId            Unique edge id (predicate___from___to)
+ * @param  {String}   params.from                    Subject node ID
+ * @param  {Array}    params.nodesIdsToDisplay       Array of nodes IDs to display
+ * @param  {String}   params.to                      Object node ID
+ * @return {Boolean}  output                         edge displaybility flag
+ */
 const showEdgeCheck = ({
   addedEdges,
-  edgeUniqueId,
   edgeId,
-  from,
-  to,
   edgesIdsToDisplay,
-  nodesIdsToDisplay
+  edgeUniqueId,
+  from,
+  nodesIdsToDisplay,
+  to,
 }) => {
   if (
     !nodesIdsToDisplay.includes(to)